+++ dist2/src/index.js
@@ -1,6 +1,7 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.sideBySideLineByLine = exports.convertTableToExcel = exports.generateExcel = void 0;
+exports.themeBaseGenerate = exports.sideBySideLineByLine = exports.convertTableToExcel = exports.generateExcel = void 0;
+const theme_1 = require("./Themes/theme");
 const create_excel_data_1 = require("./utils/create-excel-data");
 const generate_excel_1 = require("./utils/generate-excel");
 const side_by_side_1 = require("./utils/side-by-side");
@@ -15,4 +16,8 @@ function sideBySideLineByLine(data) {
     return (0, exports.generateExcel)(exData);
 }
 exports.sideBySideLineByLine = sideBySideLineByLine;
+function themeBaseGenerate(data, index, option) {
+    return (0, exports.generateExcel)((0, theme_1.themeGenerator)(data, index, option));
+}
+exports.themeBaseGenerate = themeBaseGenerate;
 //# sourceMappingURL=index.js.map