+++ src/data-model/excel-table.ts
@@ -180,7 +180,7 @@ export interface Title {
   multiStyleValue?: MultiStyleValue;
   comment?: Comment | string;
 }
-export interface Sheet {
+export interface SheetOption {
   withoutHeader?: boolean;
   multiStyleConditin?: MultiStyleConditinFunction;
   useSplitBaseOnMatch?: boolean;
@@ -202,6 +202,8 @@ export interface Sheet {
   headerRowOption?: any;
   protectionOption?: ProtectionOption;
   headerHeight?: number;
+}
+export interface Sheet extends SheetOption {
   headers: Header[];
   data: Data[];
 }
@@ -268,8 +270,10 @@ export interface FormulaSetting {
 export interface Formula {
   [insertCell: string]: FormulaSetting;
 }
-
-export interface ExcelTable {
+export interface Theme extends ExcelTableOption {
+  sheet: SheetOption[];
+}
+export interface ExcelTableOption {
   notSave?: boolean;
   creator?: string;
   backend?: boolean;
@@ -282,5 +286,7 @@ export interface ExcelTable {
   createType?: string;
   mapSheetDataOption?: any;
   styles?: Styles;
+}
+export interface ExcelTable extends ExcelTableOption{
   sheet: Sheet[];
 }